Transform Policy API to ListenForPolicyViolations
Added new api for setting policy for all the gpus instead of individual
gpus. Policy API is deprecated as the due to usability constraints.
Moreover listening to policy violations for one gpu at a time is not
very useful.
With ListenForPolicyViolations API all of the gpus are registered and
listened for policy violations instead of individual GPUs. Policy
callbacks can be registered only once during the lifetime of the program.